IGAPP-378: Create separate component for native html
