+++ native/package.json
@@ -69,7 +69,7 @@
     "react-native-pdf": "6.3.0",
     "react-native-permissions": "3.0.1",
     "react-native-reanimated": "^2.0.0",
-    "react-native-render-html": "^6.0.4",
+    "react-native-render-html": "^6.0.5",
     "react-native-safe-area-context": "^3.2.0",
     "react-native-screens": "^3.4.0",
     "react-native-static-server": "^0.5.0",
