+++ native/src/components/__tests__/NewsDetail.spec.tsx
@@ -5,6 +5,8 @@ import buildConfig from '../../constants/buildConfig'
 import React from 'react'
 import NewsDetail from '../NewsDetail'
 
+jest.mock('react-i18next')
+
 const testHTML = `<main><p>ArbeitnehmerInnen in Quarant&#228;ne haben nicht zwangsl&#228;ufig frei.</p>\n<p>tun21033101</p>\n
   <h1><a href="https://tunewsinternational.com/category/corona-deutsch/">Aktuelle Informationen zu Corona: Hier klicken</a></h1>\n</main>\n`
 const tuNews = new TunewsModel({
@@ -46,7 +48,7 @@ describe('NewsDetail', () => {
     fireEvent.press(getByText('mailto:app@integreat-app.de'))
     expect(navigateToLink).toHaveBeenCalledWith('mailto:app@integreat-app.de', language, 'mailto:app@integreat-app.de')
     fireEvent.press(getByText('https://integreat.app'))
-    expect(navigateToLink).toHaveBeenCalledWith('https://integreat.app', language, 'https://integreat.app')
+    expect(navigateToLink).toHaveBeenCalledWith('https://integreat.app/', language, 'https://integreat.app/')
   })
   it('should correctly render a tu news item', () => {
     const { getByText, queryByText } = render(
